同步更新 1.20

+++ src/main/java/com/github/tartaricacid/touhoulittlemaid/entity/task/TaskMelon.java
@@ -3,14 +3,23 @@ package com.github.tartaricacid.touhoulittlemaid.entity.task;
 import com.github.tartaricacid.touhoulittlemaid.TouhouLittleMaid;
 import com.github.tartaricacid.touhoulittlemaid.api.task.IFarmTask;
 import com.github.tartaricacid.touhoulittlemaid.entity.passive.EntityMaid;
+import com.google.common.base.Predicates;
+import com.mojang.datafixers.util.Pair;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
+import net.minecraft.enchantment.EnchantmentHelper;
+import net.minecraft.enchantment.Enchantments;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
 import net.minecraft.util.Direction;
+import net.minecraft.util.Hand;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.BlockPos;
+
+import java.util.Collections;
+import java.util.List;
+import java.util.function.Predicate;
 
 public class TaskMelon implements IFarmTask {
     public static final ResourceLocation UID = new ResourceLocation(TouhouLittleMaid.MOD_ID, "melon");
@@ -47,7 +56,15 @@ public class TaskMelon implements IFarmTask {
 
     @Override
     public void harvest(EntityMaid maid, BlockPos cropPos, BlockState cropState) {
-        maid.destroyBlock(cropPos);
+        ItemStack mainHandItem = maid.getMainHandItem();
+        if (cropState.is(Blocks.MELON) && EnchantmentHelper.getItemEnchantmentLevel(Enchantments.SILK_TOUCH, mainHandItem) > 0) {
+            if (maid.destroyBlock(cropPos, false)) {
+                mainHandItem.hurtAndBreak(1, maid, (e) -> e.broadcastBreakEvent(Hand.MAIN_HAND));
+                Block.popResource(maid.level, cropPos, Items.MELON.getDefaultInstance());
+            }
+        } else {
+            maid.destroyBlock(cropPos);
+        }
     }
 
     @Override
@@ -63,5 +80,10 @@ public class TaskMelon implements IFarmTask {
     @Override
     public double getCloseEnoughDist() {
         return 1.5;
+    }
+
+    @Override
+    public List<Pair<String, Predicate<EntityMaid>>> getConditionDescription(EntityMaid maid) {
+        return Collections.singletonList(Pair.of("has_silk_touch", Predicates.alwaysTrue()));
     }
 }
